% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_cod.R
\name{find_cod}
\alias{find_cod}
\title{Find state or municipality information}
\usage{
find_cod(query)
}
\arguments{
\item{query}{name of city/state, is a character}
}
\value{
\code{tibble}
}
\description{
Find state or municipality information
}
\examples{
find_cod("rio de janeiro")
find_cod("bahia")
}
