% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_ente.R
\name{find_ente}
\alias{find_ente}
\title{find city or state information}
\usage{
find_ente(query = NULL, v = FALSE)
}
\arguments{
\item{query}{is a character}

\item{v}{is a logical, \code{TRUE} open Data Viewer. Default is \code{FALSE}}
}
\value{
\code{tibble}
}
\description{
find city or state information
}
\examples{
find_ente("rio de janeiro")
find_ente("bahia")
}
