% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_annex.R
\name{get_annex}
\alias{get_annex}
\title{Attachments of reports by Sphere of Government}
\usage{
get_annex(verbose = FALSE)
}
\arguments{
\item{verbose}{is a logical. Enable verbose mode. Default is \code{FALSE}}
}
\value{
\code{tibble}
}
\description{
Refers to the annexes to each of the reports / statements grouped by sphere of government
}
\examples{
get_annex()
}
