% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_annual_acc.R
\name{get_annual_acc}
\alias{get_annual_acc}
\title{Annual Statement of Accounts}
\usage{
get_annual_acc(year, annex = NULL, cod, verbose = FALSE)
}
\arguments{
\item{year}{is a numeric vector}

\item{annex}{is a character vector. Default is \code{NULL},
it will get all attachments.}

\item{cod}{is a numeric vector. Brazilian Institute of Geography and
Statistics (IBGE) code assigned to each municipality and state.}

\item{verbose}{is a logical. Enable verbose mode. Default is \code{FALSE}}
}
\value{
\code{tibble}
}
\description{
Access the Annual Statement of Accounts
}
\details{
Possible values for the \code{annex} parameter are: \code{"Anexo I-AB",
"Anexo I-C", "Anexo I-D", "Anexo I-E", "Anexo I-F", "Anexo I-G",
"Anexo I-HI", "DCA-Anexo I-AB", "DCA-Anexo I-C", "DCA-Anexo I-D",
"DCA-Anexo I-E", "DCA-Anexo I-F", "DCA-Anexo I-G", "DCA-Anexo I-HI"}
For more details on the parameter \code{cod} see the column \code{cod_ibge}
of the function \code{get_info()}
}
\note{
Brazilian Public Sector Accounting and Tax Information System (Siconfi):
http://apidatalake.tesouro.gov.br/docs/siconfi/
}
\examples{
get_annual_acc(year = 2019, cod = 35)
}
