% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_info.R
\name{get_info}
\alias{get_info}
\title{Basic information of the federation entities}
\usage{
get_info(verbose = FALSE)
}
\arguments{
\item{verbose}{is a logical. Enable verbose mode. Default is \code{FALSE}}
}
\value{
\code{tibble}
}
\description{
Get basic information about all the entities of the federation
}
\note{
Brazilian Public Sector Accounting and Tax Information System (Siconfi):
http://apidatalake.tesouro.gov.br/docs/siconfi/
}
