% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_status.R
\name{report_status}
\alias{report_status}
\title{Report Status}
\usage{
report_status(year, cod, verbose = FALSE)
}
\arguments{
\item{year}{is a numeric vector}

\item{cod}{is a numeric vector. Brazilian Institute of Geography and
Statistics (IBGE) code assigned to each municipality and state.}

\item{verbose}{is a logical. Enable verbose mode. Default is \code{FALSE}}
}
\value{
\code{tibble}
}
\description{
Information on approved and corrected reports as well as 
matrices delivered to Siconfi.
}
\note{
Brazilian Public Sector Accounting and Tax Information System (Siconfi): 
http://apidatalake.tesouro.gov.br/docs/siconfi/
}
\examples{
report_status(year = 2020, cod = 35)
}
