% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_cod.R
\name{find_cod}
\alias{find_cod}
\title{Find state or municipality information}
\usage{
find_cod(query = NULL, view = FALSE)
}
\arguments{
\item{query}{is a character}

\item{view}{is a logical, \code{TRUE} open Data Viewer. Default is \code{FALSE}}
}
\value{
\code{tibble}
}
\description{
Find state or municipality information
}
\examples{
find_cod("rio de janeiro")
find_cod("bahia")
}
