% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_entes.R
\name{get_entes}
\alias{get_entes}
\title{Basic information of the federation entities}
\usage{
get_entes()
}
\value{
\code{tibble}
}
\description{
Get basic information about all the entities of the federation:
}
\note{
Brazilian Public Sector Accounting and Tax Information System (Siconfi):
http://apidatalake.tesouro.gov.br/docs/siconfi/
}
