% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rreo.R
\name{get_rreo}
\alias{get_rreo}
\title{Budget Execution Summary Report}
\usage{
get_rreo(year, period, simple = FALSE, annex = NULL, id)
}
\arguments{
\item{year}{is a numeric vector}

\item{period}{is a numeric vector, bimonth of the year, an integer between
1 and 6}

\item{simple}{is a logical. \code{TRUE} applies only to municipalities with
less than 50 thousand inhabitants that have opted for simplified publication.
Default is \code{FALSE}}

\item{annex}{is a character vector. The default value is \code{NULL},
it will get all attachments.}

\item{id}{is a numeric vector. Brazilian Institute of Geography and
Statistics (IBGE) code assigned to each municipality and state.}
}
\value{
\code{tibble}
}
\description{
Access the Budget Execution Summary Report
}
\details{
Possible values for the \code{annex} parameter are: \code{"01", "02", "03",
"04", "04 - RGPS", "04 - RPPS", "04.0 - RGPS", "04.1", "04.2",
"04.3 - RGPS", "05", "06", "07", "09", "10 - RGPS", "10 - RPPS", "11",
"13", "14"}.
For more details on the parameter \code{id} see the column \code{cod_ibge}
of the function \code{get_entes()}
}
\note{
Brazilian Public Sector Accounting and Tax Information System (Siconfi): 
http://apidatalake.tesouro.gov.br/docs/siconfi/
}
\examples{
get_rreo(year = 2020, period = 1, id = 35)
get_rreo(year = 2019:2020, period = 2, id = 35)
}
