% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siconfir-package.R
\docType{package}
\name{siconfir-package}
\alias{siconfir}
\alias{siconfir-package}
\title{siconfir: Access Brazilian Fiscal Data from SICONFI API}
\description{
\if{html}{\figure{logo.svg}{options: style='float: right' alt='logo' width='120'}}

Provides functions to access fiscal data from the Brazilian National Treasury's SICONFI API \url{https://apidatalake.tesouro.gov.br/docs/siconfi/}. Includes budget execution reports (RREO), fiscal management reports (RGF), annual accounts declarations (DCA), accounting balances matrices (MSC), and entity information for all Brazilian states and municipalities. Features automatic pagination, in-memory caching, and tidy output.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/StrategicProjects/siconfir}
  \item \url{https://strategicprojects.github.io/siconfir/}
  \item Report bugs at \url{https://github.com/StrategicProjects/siconfir/issues}
}

}
\author{
\strong{Maintainer}: André Leite \email{leite@castlab.org}

Authors:
\itemize{
  \item André Leite \email{leite@castlab.org}
  \item Marcos Wasiliew \email{marcos.wasiliew@gmail.com} (\href{https://orcid.org/0009-0004-4694-3159}{ORCID})
  \item Hugo Vasconcelos \email{hugo.vasconcelos@ufpe.br}
  \item Carlos Amorim \email{carlos.agaf@ufpe.br}
  \item Diogo Bezerra \email{diogo.bezerra@ufpe.br}
  \item Tiago Pereira \email{tiago.pereira@tesouro.gov.br}
  \item Fernando Barbalho \email{fbarbalho@gmail.com}
  \item Júlia Nascimento Barreto \email{juliabarreto@gd.seplag.pe.gov.br} (\href{https://orcid.org/0009-0004-2851-7770}{ORCID})
}

}
\keyword{internal}
