% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siconfir-package.R
\docType{package}
\name{siconfir-package}
\alias{siconfir}
\alias{siconfir-package}
\title{siconfir: Access Brazilian Fiscal Data from SICONFI API}
\description{
Provides functions to access fiscal data from the Brazilian National Treasury's SICONFI API \url{https://apidatalake.tesouro.gov.br/docs/siconfi/}. Includes budget execution reports (RREO), fiscal management reports (RGF), annual accounts declarations (DCA), accounting balances matrices (MSC), and entity information for all Brazilian states and municipalities. Features automatic pagination, in-memory caching, and tidy output.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/StrategicProjects/siconfir}
  \item \url{https://strategicprojects.github.io/siconfir/}
  \item Report bugs at \url{https://github.com/StrategicProjects/siconfir/issues}
}

}
\author{
\strong{Maintainer}: Andre Leite \email{leite@castlab.org}

Authors:
\itemize{
  \item Marcos Wasilew \email{marcos.wasilew@gmail.com}
  \item Hugo Vasconcelos \email{hugo.vasconcelos@ufpe.br}
  \item Carlos Amorin \email{carlos.agaf@ufpe.br}
  \item Diogo Bezerra \email{diogo.bezerra@ufpe.br}
  \item Tiago Pereira \email{tiago.pereira@tesouro.gov.br}
  \item Fernando Barbalho \email{fbarbalho@gmail.com}
}

}
\keyword{internal}
